% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gboot_plot.R
\name{gboot_plot}
\alias{gboot_plot}
\title{Bootstrap plot}
\usage{
gboot_plot(x)
}
\arguments{
\item{x}{object generate by functions \code{\link[geotoolsR]{gboot_block}},
\code{\link[geotoolsR]{gboot_cloud}}, \code{\link[geotoolsR]{gboot_cross}},
 \code{\link[geotoolsR]{gboot_solow}}, \code{\link[geotoolsR]{gboot_variogram}}}
}
\description{
A graphic with the original variogram and each of the B bootstrap variograms.
}
\details{
Examples of this function can be found in \code{\link[geotoolsR]{gboot_block}},
 \code{\link[geotoolsR]{gboot_cloud}}, \code{\link[geotoolsR]{gboot_cross}},
  \code{\link[geotoolsR]{gboot_solow}}, \code{\link[geotoolsR]{gboot_variogram}}
}
\author{
Diogo Francisco Rossoni \email{dfrossoni@uem.br}

Vinicius Basseto Felix \email{felix_prot@hotmail.com}
}
