% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilmoisture.R
\docType{data}
\name{soilmoisture}
\alias{soilmoisture}
\title{Soil moisture experiment}
\format{An object of class data.frame}
\usage{
data(soilmoisture)
}
\description{
Field experiment, realized in CTI (Tecnical Center of Irrigation), in an area of
350 cm x 150 cm, with 355 observations in a regular grid.
}
\details{
\itemize{
\item \bold{x} a vector containing the sample locations(in cm) in the horizontal.
\item \bold{y} a vector containing the sample locations(in cm) in the vertical.
\item \bold{z} a vector containing the value of soil moisture in percents.
}
}
\references{
HARA. A. T., GONÇALVES, A. C. A. Temporal stability of the spatial pattern of water storage in the soil at different
spatial scales. (Doctoral thesis). Retrieved from url \url{http://www.pga.uem.br/dissertacao-tese/710}
}
\keyword{datasets}
